refactor: reorganize message handlers dir
