+++ moqt-server/src/modules.rs
@@ -1,11 +1,7 @@
 pub mod buffer_manager;
-pub(crate) mod control_message_handler;
-pub(crate) mod handlers;
 pub(crate) mod logging;
+pub(crate) mod message_handlers;
 pub(crate) mod moqt_client;
 pub(crate) mod object_cache_storage;
-pub(crate) mod object_stream_handler;
 pub(crate) mod pubsub_relation_manager;
 pub(crate) mod send_stream_dispatcher;
-pub(crate) mod server_processes;
-pub(crate) mod stream_header_handler;
