+++ moqt-server/src/modules/message_handlers/control_message_handler/handlers.rs
@@ -1,8 +1,6 @@
 pub(crate) mod announce_handler;
 pub(crate) mod announce_ok_handler;
 pub(crate) mod server_setup_handler;
-pub(crate) mod stream_subgroup_header_handler;
-pub(crate) mod stream_track_header_handler;
 pub(crate) mod subscribe_error_handler;
 pub(crate) mod subscribe_handler;
 pub(crate) mod subscribe_namespace_handler;
