+++ moqt-server/src/modules/message_handlers/control_message_handler/server_processes.rs
@@ -2,8 +2,6 @@ pub(crate) mod announce_error_message;
 pub(crate) mod announce_message;
 pub(crate) mod announce_ok_message;
 pub(crate) mod client_setup_message;
-pub(crate) mod stream_track_header;
-pub(crate) mod stream_track_subgroup;
 pub(crate) mod subscribe_error_message;
 pub(crate) mod subscribe_message;
 pub(crate) mod subscribe_namespace_message;
