+++ moqt-server/src/modules/message_handlers/control_message_handler/server_processes/announce_ok_message.rs
@@ -6,7 +6,10 @@ use moqt_core::{
     pubsub_relation_manager_repository::PubSubRelationManagerRepository,
 };
 
-use crate::modules::{handlers::announce_ok_handler::announce_ok_handler, moqt_client::MOQTClient};
+use crate::modules::{
+    message_handlers::control_message_handler::handlers::announce_ok_handler::announce_ok_handler,
+    moqt_client::MOQTClient,
+};
 
 pub(crate) async fn process_announce_ok_message(
     payload_buf: &mut BytesMut,
