+++ moqt-server/src/modules/message_handlers/control_message_handler/server_processes/subscribe_message.rs
@@ -1,6 +1,7 @@
 use crate::modules::moqt_client::MOQTClient;
 use crate::modules::{
-    handlers::subscribe_handler::subscribe_handler, object_cache_storage::ObjectCacheStorageWrapper,
+    message_handlers::control_message_handler::handlers::subscribe_handler::subscribe_handler,
+    object_cache_storage::ObjectCacheStorageWrapper,
 };
 use crate::SenderToOpenSubscription;
 use anyhow::{bail, Result};
