+++ moqt-server/src/modules/message_handlers/control_message_handler/server_processes/subscribe_ok_message.rs
@@ -8,7 +8,8 @@ use moqt_core::{
 };
 
 use crate::modules::{
-    handlers::subscribe_ok_handler::subscribe_ok_handler, moqt_client::MOQTClient,
+    message_handlers::control_message_handler::handlers::subscribe_ok_handler::subscribe_ok_handler,
+    moqt_client::MOQTClient,
 };
 
 pub(crate) async fn process_subscribe_ok_message(
